+++ plotoptix/enums.py
@@ -305,8 +305,30 @@ class Geometry(Enum):
     interpolating all data points.
     """
 
+    BSplineQuadRocaps = 22
+    """Quadratic b-spline with nodes at data points (roving capsules implementation).
+
+    Curve thickness and color can be provided for each data point (curve node).
+    
+    Note: b-spline is not interpolating data points; see examples
+    how to pin start/end to a fixed position. Use
+    :attr:`plotoptix.enums.Geometry.BezierChain` for a smooth curve
+    interpolating all data points.
+    """
+
     BSplineCubic = 13
     """Cubic b-spline with nodes at data points.
+
+    Curve thickness and color can be provided for each data point (curve node).
+    
+    Note: b-spline is not interpolating data points; see examples
+    how to pin start/end to a fixed position. Use
+    :attr:`plotoptix.enums.Geometry.BezierChain` for a smooth curve
+    interpolating all data points.
+    """
+
+    BSplineCubicRocaps = 23
+    """Cubic b-spline with nodes at data points (roving capsules implementation).
 
     Curve thickness and color can be provided for each data point (curve node).
     
@@ -323,8 +345,26 @@ class Geometry(Enum):
     Curve interpolates its data points (nodes) exactly.
     """
 
+    CatmullRomRocaps = 24
+    """Catmull-Rom spline with nodes at data points (roving capsules implementation).
+
+    Curve thickness and color can be provided for each data point (curve node).
+    Curve interpolates its data points (nodes) exactly.
+    """
+
     Beziers = 16
     """Bezier curves with data points used as **control points** for each independent segment (OptiX native implementation).
+
+    Curve thickness and color can be provided for each data point (bezier control point).
+    Each segment is described with 4 control points: start, ctrl #1, ctrl #2, end. Multiple
+    segments are described with an array of control point 3D positions (colors, radii) of
+    successive segments, i.g. N segments will require an array ``(4*N,3)`` of positions,
+    an array ``(4*N,3)`` of colors and an array ``(N,)`` of radii, see also ``2_beziers.py``
+    in basic code examples.
+    """
+
+    BezierRocaps = 25
+    """Bezier curves with data points used as **control points** for each independent segment (roving capsules, OptiX native implementation).
 
     Curve thickness and color can be provided for each data point (bezier control point).
     Each segment is described with 4 control points: start, ctrl #1, ctrl #2, end. Multiple